Change some scripts: speed up |theta_stable_parabolics|, some general cleaning up
While moving some definitions from misc.at into basic.at, it was decided that
"distinguished fiber" is to preferred to "fundamental fiber" (which the Atlas
source code attests throughout, no doubt going back to Fokko) for the set of KGB
elements whose involution is the distinguished involution of the inner class.

Nonetheless there is only a function |distinguished_fiber@RealForm|, which picks
out the proper part of the KGB set. To match the term "fundamental fiber" in the
Atlas source code, one should group together elements of different real forms,
even of different strong real forms giving isomorphic real forms, as in the
|fiber_partition| for a given |Cartan_class|. This would require a bit of work,
and there does not currently seem to be any need for it.