Restructure big_unitary_hash...
...so that instead of saving separate unitary and nonunitary hashes,
it saves one Param_hash and an [int] giving statuses of each
Param. Since Uhash was pervasive in to_ht_branch, this required a
great many small changes.

Plan is that the new structure should make it easier to use
deformation arguments for unitarity; nothing is yet changed in that
direction. New code seems to work OK, but it's a little slower than
to_ht_branch for now.