Fix and modify generation of type E conjugacy class tables and character tables
For E8, the table for the power map in the |WeylClassTable| construction was
given in terms of the output of |conjugacy_class_reps@RootDatum|, which had
changed, so that the power map was no longer correct. This commit fixes the
table, now relative to an ordering of the conjugacy classes lexicographically by
root orbit type. We also now provide as a comment the code that could recreate
the table dynamically, making it even more robust against permutations of the
conjugacy class listing (though we think the currently chosen ordering is
canonical and natural, and therefore probably will not have to change). As a
side effect of this modification, we adapt the order produced by the
|class_table_E8| functions to be this canonical order, knowing that the
character table construction was already robust against any permutations of the
conjugacy classes in the |WeylClassTable| presented to it.

We similarly adapted the construction of the type E6 and E7 conjugacy class and
character tables so as to present similar properties, even though these
constructions were not previously broken.