Adapt type F and G conjugacy class and character table constructions like type E
This commit follows suit of the previous commit, so that for all exceptional
types the conjugacy classes are by default ordered by signature (essentially the
type of the partition of the set of roots into orbits) lexicographically.

For type G2, this involved editing the numbers used during construction of the
table of conjugacy classes, which is delicate but seems to have worked.