Change type EFG character table constructors to retain conjugacy class ordering
The constructors of character tables for (simple) exceptional types now retain
the ordering of conjugacy classes used in the WeylClassTable passed to them,
rather than change to the convention used by by the external tools that have
produced the character table data. This means that each character is permuted as
needed to match the class table. This intermediate stage of constructing from a
class table rather than directly from a root datum allows an explicit ordering
to be imposed on the table using the |renumber| function before giving the
result to the character table constructor, if a non standard order is desired.

permute the lines of those character tables while building our own tables).