Too many small changes...
One is that the use of the Dirac inequality in conjunction with making
families of reps having the same signature to HT was mathematically
unjustified. Failure of Dirac on tau means that the signature is
indefinite on tau and SOME KType tau' appearing in tau\otimes
(g/k). So that indefiniteness might not appear up to HT; so failure of
Dirac on one rep in a family does not prove the whole family is
nonunitary.

So added to Dirac.at a function tensor_p@KType estimating the
components of tau\otimes (g/k). Used their heights to modify the
best_Dirac test accordingly, so its application to these families is
now justified. (The old unjustified behavior can be recovered with
old_best_flag:=true; default is false.)

I THINK that the original Dirac test (involving just the LKT) is
applied only to individual reps, not to families, so it's OK. But
eventually I'll look harder.

Added a function grow(int N) to Hash; this enlarges the pool so that N
additional members can be accommodated without rehashing. This
provides maybe a factor of two speedup in loading large lists into a
hash table; noticeable for N around a million or more??

The automatic setting of more_shift in FPP_globalDirac is kind of a
mess. I think the defaults are pretty good for E7_s, and I tried to
make them not terrible for E8_s and E8_q; but probably it's best to do

test_bl_flag:=false

which will then use whatever more_shift(?) you use, and will always
use bottom_layer when asked.

Plan when I started this a few days ago was to use Dirac inequality in
steve.at. I did use the basic (LKT) Dirac inequality, still  being
careless about HT (fix soon!). But I did not add best_Dirac to steve
yet, because I wanted to fix the math error first.

Other changes were output tweaks and I forget what else; nothing else
big.