Mathematically cleaned Dirac.at
Somewhat improved/corrected the tensor_p functions from the last
commit. They are still too slow for large G (certainly a problem in
E7_s); I'll work on that.

The adjust_Dirac function decides whether it's a good idea to use
Dirac and/or DiracBest. Now this is done only if adjust_Dirac_flag is
true (the default); and the functions decision to turn on Dirac does
NOT override user's decision not to with Dirac_flag or Dirac_best_flag
(both still defaulting to true).