Fiddled with the code decided whether to use bottom_layer.
This is now controlled by two flags: easy_bl_flag and hard_bl_flag,
defaulting to true and false. If hard_bl_flag is true, then the test
is a little slow (probably not significant for large (x,lambda?) but
maybe better at picking a good more_shift. If hard_bl_flag is false
and easy_bl_flag is true, then there is a reasonable decision about
using bottom_layer, but more_shift is more haphazard. If both are
false, then the use of bottom_layer and the more_shift level are as
specified by the user.

Perhaps because of the extra work in Dirac inequality, this version is
maybe 5-10% slower than best previous at the beginning of E7_s. We'll
see if it catches up.