Fixed fix of Dirac inequality
Previously I added the capability of correctly bounding the height
where Dirac guarantees a signature change; but I think that I failed
to make the code use the capability. (The computation is not quite
trivial, so I put a placeholder -1 if I didn't need this bound, with
the idea of modifying it to the correct bound when asked. But then I
never asked.

I believe it is now fixed. Through KGB(E7_s,10501), the last commit
found 12387 nonunitary Params with min_bottom. The present code finds
12426. I believe that the 39 extras found now were (unjustifiably)
excluded by Dirac in the previous commit. (I will try to verify that.)
Both versions found in the end exactly the same unitary
representations through KGB(E7_s,10501).