improved is_unitary_to_hts_big_SIMPLE_interrupt...
...by increasing the time allotted to a test ONLY when the previous
attempt made no progress (failed to record an intermediate
full_deform). On Jeff's bad example

(KGB(E8_q, 66989),[ 4, 0, 0, 1, 1, 1, 1, 2 ]/1)

this reduced the time on my laptop from 9300 seconds to 3300 seconds.

Consequence is that all these functions now ASSUME that
deformed_hash_flag is true. If it is not, then the functions will
offer more and more time to all interruptible tests; this will cost
some time, but is not fatal. Easy to fix.