Rewrite permutation code in combinatorics.at, with |Permutation=vec|, not |[int]|
We use |set_type Permutation = vec|, with a predicate |is_permutation|, for
anything involving pure permutations, for readability and efficiency. We still
use |[int]| for (possibly) signed permutations though.

Some code was slightly rewritten, and |permutation_conjugate| was repaired (and
spelling-corrected); the permutation code was also moved to what seems a more
proper place in the file (and where it used to be? it is not clear to me).

Some improvements were also made to diagram.at, currently the only user of
permutation iterators, which are now initialized using |vec| rather than |int|.
This is actually used now to represent permutations of a cluster of equal simple
types in a |LieType| as permutations of the set of indices of those simple types
rather than of a consecutive range of integers.