clean up for speed
The use of u_predictor in almost all the is_unitary commands changes
their behavior significantly. One result is that my old testing
(suggesting that edge_offset = 3 was ideal) is no longer applicable:
the default is now edge_offset = 1. (Probably a larger value is needed
in the spherical case???)

The reversion to steve should I think happen only for split G and only
in the non-spherical cases on x_open. I think this is now
done. Formerly (just sloppiness) the spherical case also reverted to
steve when the argument was [(KGBElt,ratvec)]; this is now fixed.

steve.at has been modified to make better use of the modified
is_unitary commands; haven't worked hard on that, but it seems closer
to reasonable now.