worked on next_heightC(Param p, int HT)
This returns the height of the first K-type of p of height > HT, or -1
if none exist. The previous version worked, but was extremely slow in
E8_q because of computing K_type_formula. This version only computes
K_type_formula to some height. I think it's arranged to work in
reasonable time for E8_q, and I'm pretty sure it's arranged always to
give a good answer.

Use a new function tensor_p_big(KType mu) = KType, which finds the
constituent of mu\otimes(g/k) of highest height.