Fix error in logic of hashing deformation unit in case of twisted deformation
When a previously computed twisted deformation is used, the test for whether a
flip is needed with respect to the default extension used when it was stored
requires the stored sample to be delta-fixed (and correspond to the twisted
terms stored). In order for this to work, the stored |sample| parameter needs to
be updated when a twisted deformation is first requested in an alcove that
already has been visted for an untwisted deformation. That update is now done.