Added big_unitary_hash.writeG_nu(RealForm G)... ...which when point to a file, produces an .at file which (on loading) records in big_unitary_hash all the nonunitary reps known to (the old) big_unitary_hash.
This reduces but does not eliminate nonunitarity results. It can
happen that big_unitary_hash only knows p is nonunitary, but not a
height to which it is known. In this case it's very common for
is_unitary_to_ht(p) to be recomputed.

Moral is that I should try harder to record these known heights.

New function status(p,HT) returns 1 if big_unitary_hash KNOWS
that is_unitary_to_ht(p,HT) will return true, -1 if it knows false,
and 0 otherwise.