got rid of a few function and variable re-definitions.
None seemed to matter; often just repeating a definition from an
earlier file or even an earlier line in the same file.

There was some problem with FPP_unitary_hash_bottom_layer aborting
before KGBElt 16000 because of assigning to a non-existent entry
in was_tested. First instance of this (few days ago) was an obvious
scripting error. This time I'm not clear what happened, so not certain
whether it's fixed. Testing takes five hours; we'll see...