Two serious errors...
In big_unitary_hash, the function

 sig_entry_interrupt

was defined to add its argument to phash BEFORE seeing whether the
computation would time out. If this happened, then phash no longer
matched the rest of big_unitary_hash. I don't know what errors this
caused (one possibility is an interrupt for trying to assign to a
non-existent entry; that one I think happened to me) but the
possibilities are awful.

In steve_hash_small, line (approximately line 1350) the script tested
for nonunitarity by status() > 0; should have been < 0. The result was
failing to find hundreds of unitary candidates. If E7_s, a single one
of these

plost = parameter(E7_s,20925,[2,1,1,2,1,2,1],[2,1,2,1,1,1,1]/6)

actually WAS unitary, and so was missed by FPP_unitary_hash* when
using steve on x_open. Result was an FPP_unitary_dual of size 237640
instead of 237641. This error took me several days to find, and none
of them were fun.

While looking for this second error, I fiddled with a lot of code;
nothing should be substantive.